+++ sensei/__init__.py
@@ -1,6 +1,6 @@
+from ._internal import Path, Query, Cookie, Header, Body, File, Form
 from ._internal import Router, Args, APIModel
 from ._utils import format_str, placeholders
 from .cases import *
 from .client import RateLimit, Manager, Client, AsyncClient
-from .params_functions import Path, Query, Cookie, Header, Body, File, Form
 from .types import Json
