+++ sensei/_internal/_core/api_model.py
@@ -1,34 +1,48 @@
 import functools
-from typing import Any
+from typing import Any, Callable
 
 from pydantic import BaseModel
 from pydantic._internal._model_construction import ModelMetaclass
+from typing_extensions import TypeGuard
 
 from sensei.types import Json
-from ._endpoint import Args
-from ._hook import Hook
-from ._types import RoutedMethod
-from ..tools import is_routed_method
-from ..tools import is_staticmethod, is_classmethod, is_selfmethod, bind_attributes
+from ._types import RoutedMethod, ModelHook
+from .args import Args
+from ..tools import is_staticmethod, is_classmethod, is_selfmethod, bind_attributes, is_method
 
 
 class _Namespace(dict):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.__hooks: dict[str, Callable] = {}
+        self._routed_keys = set()
+
     @staticmethod
     def _decorate_method(method: RoutedMethod) -> None:
-        try:
-            preparer = method.__func__.prepare
-            finalizer = method.__func__.finalize
+        preparer = method.__func__.prepare
+        finalizer = method.__func__.finalize
+        method.__route__ = method.__func__.__route__
 
-            bind_attributes(method, finalizer, preparer)  # type: ignore
-        except AttributeError:
-            pass
+        bind_attributes(method, finalizer, preparer)  # type: ignore
+
+    @staticmethod
+    def _is_routed_method(obj: Any) -> TypeGuard[RoutedMethod]:
+        cond = False
+        if is_method(obj):
+            func = obj.__func__
+            cond = getattr(func, '__routed__', None) is True
+        return cond
 
     def __setitem__(self, key: Any, value: Any):
         if is_staticmethod(value) or is_classmethod(value):
-            if is_routed_method(value):
+            if self._is_routed_method(value):
                 self._decorate_method(value)
-        elif key in Hook.values() and is_selfmethod(value):
-            value = functools.partial(value, None)
+                value.__route__.hooks.set_model_hooks(self.__hooks)
+        elif key in ModelHook.values():
+            if is_selfmethod(value):
+                value = functools.partial(value, None)
+
+            self.__hooks[key] = value
 
         super().__setitem__(key, value)
 
@@ -36,29 +50,40 @@ class _Namespace(dict):
 class _ModelMeta(ModelMetaclass):
     @classmethod
     def __prepare__(metacls, name, bases):
-        return _Namespace()
+        namespace = _Namespace()
+        return namespace
 
     def __new__(
             cls,
             cls_name: str,
             bases: tuple[type[Any], ...],
-            namespace: dict[str, Any],
+            namespace: _Namespace,
     ):
+        print('hello new')
         obj = super().__new__(cls, cls_name, bases, namespace)
         obj.__router__ = None
 
         return obj
 
+    def __init__(
+            cls,
+            cls_name: str,
+            bases: tuple[type[Any], ...],
+            namespace: _Namespace,
+    ):
+        print('hello init')
+        super().__init__(cls_name, bases, namespace)
+
 
 class APIModel(BaseModel, metaclass=_ModelMeta):
     """
-    Base class for creating Sensei API models. Don`t confuse it with Pydantic BaseModel, it`s used simultaneously
-    for validating output data and provide "routed" functions.
+    Base class for creating Sensei API models. Don't confuse it with Pydantic BaseModel, it's used simultaneously
+    for validating output data and provides "routed" functions.
 
-    But usage rools are the same with BaseModel
+    But usage rools is the same with BaseModel
     Usage docs: https://docs.pydantic.dev/2.9/concepts/models/
 
-    To make the proper class, decorate it with @router.model. Avoiding this requirements will lead you to the issues.
+    To make the proper class, decorate it with @router.model. Avoiding this requirement will lead you to the issues.
 
     Examples:
         >>> from typing import Annotated, Any, Self
@@ -66,8 +91,7 @@ class APIModel(BaseModel, metaclass=_ModelMeta):
         ...
         >>> router = Router('https://reqres.in/api')
         ...
-        >>> @router.model()
-        ... class User(APIModel):
+        >>> class User(APIModel):
         ...     email: str
         ...     id: int
         ...     first_name: str
@@ -110,6 +134,21 @@ class APIModel(BaseModel, metaclass=_ModelMeta):
             Args: The prepared arguments.
         """
         return args
+
+    @staticmethod
+    def __default_case__(s: str) -> str:
+        """
+        Convert the case of all parameters.
+
+        This hook is used to convert the case of all parameters to the desired format.
+
+        Args:
+            s (str): The original parameter string.
+
+        Returns:
+            str: The converted parameter string.
+        """
+        return s
 
     @staticmethod
     def __query_case__(s: str) -> str:
