+++ sensei/_internal/tools/__init__.py
@@ -1,4 +1,4 @@
 from .chained_map import ChainedMap
 from .types import HTTPMethod, MethodType
 from .utils import (make_model, split_params, accept_body, validate_method, args_to_kwargs, set_method_type, identical,
-                    is_staticmethod, is_classmethod, is_selfmethod, bind_attributes, is_routed_method)
+                    is_staticmethod, is_classmethod, is_selfmethod, bind_attributes, is_method)
